feat: prevent db access to dev dbs in tests
